tidy internal use statements
